+++ src/main/java/com/getpcpanel/cpp/windows/SndCtrlWindows.java
@@ -136,7 +136,7 @@ public class SndCtrlWindows implements ISndCtrl {
             StreamEx.ofValues(devices)
                     .filter(d -> "*".equals(device) || deviceId.equals(d.id()))
                     .flatCollection(d -> d.getSessions().values())
-                    .filter(s -> (fileName.equals(AudioSession.SYSTEM) && s.pid() == 0) || (s.executable() != null && StringUtils.equals(fileName, s.executable().getName())))
+                    .filter(s -> (fileName.equals(AudioSession.SYSTEM) && s.isSystemSounds()) || (s.executable() != null && StringUtils.equals(fileName, s.executable().getName())))
                     .forEach(s -> setProcessVolume(s, volume));
         }
     }